options: use interface{}, check that they exist?
if you mark peerswap as `important-plugin` and then start, it'll crash
since one of the peerswap options is not a string.

we convert to using `interface{}` as the type for Options and then
'parse' them to types inline.

while we're at it, make sure all the things exist before attempting to
use them.