tests: Refactor integration tests Co-authored-by: Peter Neuroth <nepet@users.noreply.github.com>
test: Update tests to work with premuim

As we use a premium for a swap out now, we need do update the tests
to work with the premium. The premium is calculated by the difference
of the channel balance before and after the premium invoice payment.

test: Set shorter PAYMENT_RETRY_TIME

As we have a shorter repay interval for tests now we can lower this
time to fasten up the tests.

test: Add parallelism

Add the golang flag to enable for parallel tests. This might
introduce flakyness but helps for faster local tests.