testframework: Add assertions
testframework: Set better timeout

The global TIMEOUT variable is set to a lower value as we needed
to wait for too long. With SLOW_MACHINE one can bump up the
TIMEOUT.

testframework: Add assertions for balance change

We sometimes just need to wait for a change in the channel balance.
We can use this to check if an invoice was payed.

testframework: Add logs at the end of test

This adds logs that are only presented if the test fails. This allows
using -v on go test to show the progress beeing quiet on logs from
daemons on unfailed tests.

This also adds a better filter with the filter method Filter(regex)

testframework: Add Stop method to c-lightning node

This new method does not tear down the directories.