swap: Make SetStartingBlockHeightAction resilient to restart
On restart we have to ensure, that we coop cancel the swap if we
already are outside our csv safety window.

swap: Make AwaitTxConfirmationAction resilient to restart

On restart we have to ensure, that we coop cancel the swap if we
already are outside our csv safety window.

mod: Update glightning version

test: Add Stop function to peerswap client

test: Add csv claim tests

This adds a test that triggers csv claim and checks for right amts.

fixup! test: Add csv claim tests