testframwork: Add helper methods
ci: Set SLOW_MACHINE

This allows for later timeouts such that the tests can wait longer.

testframework: Add port map

We sometimes have the problem that we assign the same port multiple
times. This map ensures, that for the time of a test runner, we only
assign each port once. This does not work over multiple runners.

testframwork: Add tail for daemon log

We want to get rid of the noise in the daemons logs. Therefore we
provide a Tail() command that works just like the tail command in
unix systems.

testframework: Add invoice methods to lightning node interface

We need methods to create and pay for invoices in the interface to
allow for more abstract tests.