lnd: Add blocking call that waits for grpc to be ready
We encountered numerous problems and errors that are the result
of a broken connection to the lnd node. This might be due to the
lnd node failing while peerswap is still alive.

This is a helper function to block on a broken grpc connection
until the connection state is READY again. This can be used to
resubscribe to a broken stream once the rpc is READY again.

The tests contain an example service that is using waitForReady
to block until a dropped connection is available again and then
resubscribes to the stream in TestResubscribeToStream.