lnd: Add retry on EOF error to grpc streams.
When lnd is shutdown gracefully some - not all - streams return an
EOF error indicating that lnd closes the connection. This is
correct grpc behavior but still we want to try to reconnect in this
case.
We Add the grpc-retry option to also retry on EOF errors to the
streams and tests to verify that the streams reconnect if lnd is
gracefully shutdown and rerun.
Also the chain notifyer returns a UNKNOWN error if lnd is shutdown
with a message that is also added to the retry ruleset.