policy: Add minimum swap amount
We already use a lower boundary in sat to perform a swap due to
economical reasons. We want to be able to return this value via the
api, so this commit moves this value to the policy where it belongs.
The MinSwapAmountMsat is now in msat, set to a default of
100000000 msat and can NOT be overridden in the policy file.
We might change this in the future.

This might change at some point in the future.