proto: Add a json marshaler to the policy message
When we generate the stubs from the proto the structs gets tagged
with json:"omitempty". When we pass this message to a json
marshaler for e.g. printing to stdout, the marshaler omits empty
fields or fields that contain the default value for their type.
What we want is to print all fields even if they are empty. This
adds a custom marshaler that is configured to emit unpopulated
values.