swap: Use the policy to detrmine if a new swap is allowed
We now use the policy instead of the inmemory map to determine if
we can start a new swap or not.

In the case of a incoming swap request from a peer, this check is
moved to the CheckRequestWrapperAction as we also want to send our
peer a cancel message if we do not allow for swaps atm.

Prior this commit the check was performed in the service methods
and did not send a cancel message to the peer so that the peer
could get stuck.