gomod: Update to lnd v0.15.2-beta
We need a fast update to the newest lnd version as there was a
security issue detected.
https://github.com/lightningnetwork/lnd/issues/7002

As the bug was part of the btcd dependencies the fix applied in
https://github.com/btcsuite/btcd/pull/1896 brings a long tail
of dependency updates.

This commit updates the dependencies and adapts for all changes
that are introduced by this new dependencies.