cln: Add wrapper for plugin start to support logs for errors
We want to gather as much infos as possible in the cln logs. Some
errors (all that where thrown in the main run routine) could not be
passed to the cln log as an error return leaves the scope in which
the cln plugin and the json rpc are available.
This commit drags the json rpc interface to an outer wrapper such
that we can pass error logs to cln from the main run routine.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>