cln: Remove peerswap-db-path option
We use a fixed db location depending on the lighting-dir. We don't
need to specify an option that points to the db location.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>

cln: Add `PeerswapDir` to config

PeerswapDir or `peerswap-dir` is the directory that contains all
peerswap relevant files: config, database, policy. It defaults to
`<lightning-dir>/peerswap`.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>

cln: Replace hard coded db name

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>