txwatcher: Rework the cln-txwatcher for recovery
Swaps that have been claimed by csv can get stuck on the peers side on
startup. This commit reworks the tx watcher to check all blocks since
the block the swaps started for the desired tx. If the tx is above the
safety limit we return an error and fail the swap.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>