test: the amount of the fee invoice is taken into account when testing
Claim coop testing of swap out time is dominating.
The reason seems to be that the channel taker balance check is stuck in an infinite loop until time out.

In the case of swap out, it seems that the amount of the fee invoice needs to be taken into account.