testframework: Fix invoice description flake
The GetFeeInvoiceAmtSat command looked for invoices with a description
that contained "fee". The invoice description always contains the
swapid which is 32 random bytes (we should use something better here
anyways) that are hex encoded. This could produce collisions where "fee"
was part of the hex encoded swapid.

We now check on a regex of the form "peerswap <ABC> fee <XYZ>" to avoid
collisions with the swapid.

Signed-off-by: Peter Neuroth <pet.v.ne@gmail.com>