lnd: Sanity check max htlc amount msat
Added logic to check the maximum htlc limit
set per node for each channel.

In cln, max htlc amt is only taken into account when foward,
so checks are only performed when the sender is lnd.
Receivable amount depends on whether the destination
node is a cln or a lnd, so max htlc amt is not taken into account.

Also, since the max htlc limit is not always set reliably,
the check is skipped if it is not set.