swap: probe payment as sanity check
Replace sanity check with probe payment in SpendableMsat.
ProbePayment trying to pay via a route with
 a random payment hash that the receiver doesn't have the preimage of.
The receiver node aren't able to settle the payment.
When the probe is successful, the receiver will return
a incorrect_or_unknown_payment_details error to the sender.
If the result of ProbePayment is assumed
to be a payment failure, log the reason as known by ProbePayment