lnd: use the SendPaymentV2 RPC and set payment_request
Use the SendPaymentV2 RPC and set payment_request in the request.
This will ensure that the payments are
linked to the peerswap invoices that were paid.
I check that the remote pubkey of the channel
and the destination of the invoice match.

Also added a test case to ensure that payreq is included in the payment.