lnd: label the tx to make it easier to audit
Ensure that all lnd transactions related to peerswaps are identifiable
by the label `peerswap`.
This makes it easier to audit the transactions from faraday.
This is performed by LND's LabelTransaction RPC.
Since CLN has no such function, no label is assigned.
Since label recording is not an essential requirement for swap,
if it fails, it is logged and subsequent processing is continued.