multi: refactor interfaces onchain wallet
`CreateFundedTransaction` and `FinalizeFundedTransaction`
have merged interfaces, CreateAndBroadcastTransaction.
This is because there was originally no reason to separate them,
and because lwk uses the psbt format,
transacrion conversion needs to be performed.

The liquid wallet interface should not have a
unique liquid wallet-dependent definition.
However, there were some that depended on gelements,
 so they were separated out.