test: set min-emergency-msat for stuck channel
Anchor channels are now the default in >24.02.
We need to account for the min-emergency-msat
now when opening a channel.

In the stuck channel test, the default
min-emergency-msat of 25000sat is too large,
so it is set to a smaller value slightly
exceeding the dust limit.

Also, in >24.02, GetRoute does not seem to
return route in stuck channel state,
so the test case has been modified.