liquid: drop the flat fee for mempool policy patch
elements releases have the mempool policy patch.
To apply ct discount on peerswap, we uses An
"Ask forgiveness not permission" approach,
where we attempt to broadcast with the discounted fee,
and if that fails, we retry with the non-discounted fee.

The discount is achieved by calculating the fee
based on a discounted vsize
(equivalent to 1/4 of the original tx size).