test: shorten socket path & improve observability
The integration tests often fail when waiting for the CLN node to
log  "Server started with public key".

Therefore, I fixed the observability and corrected the path,
as the issue was probably caused by
 the Unix domain socket being too long.
Unix domain sockets are limited to 108 bytes.
The former path (~115 chars)
The new scheme generates a much shorter path (~46 chars)

/tmp/ps/<pid>-<rand>/c1/regtest/lightning-rpc